вынос всей работы с фаербейз в один класс
